Added: module template.js
